backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
transformers
torch
slack_sdk
PyGithub
py-trello